server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "async-lru>=2.0.4",
    "fastapi>=0.115.6",
    "openai>=1.58.1",
    "peewee>=3.17.8",
    "pillow>=11.0.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
    "pytest-cov>=6.0.0",
    "uvicorn>=0.34.0",
]
